+++ scripts/nif/niftoaster.py
@@ -116,11 +116,13 @@ class NifToaster(pyffi.spells.nif.NifToaster):
         pyffi.spells.nif.modify.SpellSubstituteTexturePath,
         pyffi.spells.nif.fix.SpellDelUnusedRoots,
         pyffi.spells.nif.modify.SpellChangeBonePriorities,
+        pyffi.spells.nif.modify.SpellChangeAllBonePriorities,
         pyffi.spells.nif.modify.SpellSetInterpolatorTransRotScale,
         pyffi.spells.nif.modify.SpellDelInterpolatorTransformData,
         pyffi.spells.nif.modify.SpellCollisionToMopp,
         pyffi.spells.nif.optimize.SpellReduceGeometry,
-        pyffi.spells.nif.optimize.SpellPackCollision
+        pyffi.spells.nif.optimize.SpellOptimizeCollisionGeometry,
+        pyffi.spells.nif.optimize.SpellOptimizeAnimation
         ]
     ALIASDICT = {
         "texdump": "dump_tex",